+++ plugin/src/main/groovy/org/openbakery/carthage/CarthagePluginExtension.java
@@ -1,0 +1,12 @@
+package org.openbakery.carthage;
+
+import org.gradle.api.Project;
+
+public class CarthagePluginExtension {
+
+	boolean cache = true;
+
+	public CarthagePluginExtension(Project project) {
+	}
+
+}